Sorry big commit again(but most stuff is model related)
fixed lots of stuff in the models.
the whole top wall on the stattion turn transparent when u r behind it.
(I know I shouldn't but I also changed some small stuff in the code for wall transparency (the raycast is not to a point. it's to the camera transform) and also when the transparency effect is off I changed the material to "cutout" and when it's on the material is "fade".
(I was having a weird effect when the material was set to "fade"all the time... specially when there was some sort of intersection between walls)