Talk to Kmi at the bar!
The collider I had to put in was a bit wonky or the bar in front of her is being defficult so you have to go behind her to talk to her. Also the convo doesnt reach its real end cause I didn't put in the pet name thing. I think I MIGHT know how to go about it though!